flask/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# cli\cli.py: 1
# cli\commands\cmd_db.py: 1
# cli\commands\cmd_scripts.py: 1
# cli\commands\cmd_tasks.py: 1
Click == 7.0

# quickslack\app.py: 1
# quickslack\routes\api\slackevents.py: 1,2
# quickslack\routes\dashboard\views.py: 3,4
Flask == 1.1.1

# quickslack\app.py: 19
PyYAML == 5.3

# config\flask.py: 3
# quickslack\app.py: 2
celery == 4.4.0

# quickslack\extensions.py: 1
flask-debugtoolbar == 0.10.1

# quickslack\extensions.py: 2
flask-sqlalchemy == 2.4.1

# quickslack\routes\dashboard\forms.py: 1
flask-wtf == 0.14.2

# config\flask.py: 1
# quickslack\app.py: 3
# quickslack\sentry.py: 2
python-decouple == 3.3

# quickslack\routes\dashboard\views.py: 1
# scripts\api_test.py: 1
requests == 2.22.0

# setup.py: 1
setuptools == 41.4.0

# quickslack\routes\dashboard\forms.py: 4
wtforms == 2.2.1

# quickslack\routes\dashboard\forms.py: 2
wtforms-components == 0.10.4

# not found by pigar
gunicorn
psycopg2-binary
redis
flower

Flask-Migrate == 2.5.2
git+https://github.com/quickslack/slack-user-client
sentry-sdk[flask]==0.14.1